Move trackEvents.js to client directory
